Importer improvements
- Only call completion handler when all imports are done when passing multiple paths
- Property queue up operations in serial queue after bulk generation, with completion handler when all tasks done
 - If a specific sysem is passed to rescan, try to match existing ROMs much earlier for faster scan performance
- When re-scanning ROMs in system folders, properly match roms that are assocaited files of an already existing PVGame entry
- If can’t delete associated images, files, other than main rom file, still remove PVGame from database
- Fixed can’t delete PVGame entry if assoaciated file is missing

+++ Provenance/Game Library/Database/RomDatabase.swift
@@ -400,11 +400,13 @@ public extension RomDatabase {
 			}
 		}
 
-		do {
-			try FileManager.default.removeItem(at: romURL)
-		} catch {
-			ELOG("Unable to delete rom at path: \(romURL.path) because: \(error.localizedDescription)")
-			throw error
+		if !game.file.missing {
+			do {
+				try FileManager.default.removeItem(at: romURL)
+			} catch {
+				ELOG("Unable to delete rom at path: \(romURL.path) because: \(error.localizedDescription)")
+				throw error
+			}
 		}
 
 		// Delete from Spotlight search
@@ -414,12 +416,18 @@ public extension RomDatabase {
 		}
 		#endif
 
-		game.saveStates.forEach { try? $0.delete() }
-		game.recentPlays.forEach { try? $0.delete() }
-		game.screenShots.forEach { try? $0.delete() }
+		do {
+			game.saveStates.forEach { try? $0.delete() }
+			game.recentPlays.forEach { try? $0.delete() }
+			game.screenShots.forEach { try? $0.delete() }
 
-		try deleteRelatedFilesGame(game)
-		try game.delete()
+			try deleteRelatedFilesGame(game)
+			try game.delete()
+		} catch {
+			// Delete the DB entry anyway if any of the above files couldn't be removed
+			try game.delete()
+			throw error
+		}
 	}
 
 	func deleteRelatedFilesGame(_ game: PVGame) throws {
